# docs/api/swagger.yaml
basePath: /api
definitions:
...
  internal_modules_setting.CreateRequest:
    properties:
      description:
        maxLength: 255
        type: string
      setting_key:
        maxLength: 30
        minLength: 2
        type: string
      setting_values:
        type: string
      store_id:
        type: integer
    required:
    - setting_key
    - setting_values
    type: object
  internal_modules_setting.DefaultCreateRequest:
    properties:
      description:
        maxLength: 255
        type: string
      setting_key:
        maxLength: 30
        minLength: 2
        type: string
      setting_values:
        type: string
    required:
    - setting_key
    - setting_values
    type: object
  internal_modules_setting.DefaultDeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  internal_modules_setting.DefaultListRequest:
    properties:
      limit:
        maximum: 1000
        minimum: 1
        type: integer
      page:
        minimum: 1
        type: integer
      setting_key:
        type: string
    required:
    - limit
    - page
    type: object
  internal_modules_setting.DefaultUpdateRequest:
    properties:
      description:
        maxLength: 255
        type: string
      id:
        type: integer
      setting_key:
        maxLength: 30
        minLength: 2
        type: string
      setting_values:
        type: string
    required:
    - id
    - setting_key
    - setting_values
    type: object
  internal_modules_setting.DeleteRequest:
    properties:
      id:
        type: integer
      store_id:
        type: integer
    required:
    - id
    type: object
  internal_modules_setting.FormConfigsRequest:
    properties:
      only_platform:
        type: boolean
    type: object
  internal_modules_setting.ListRequest:
    properties:
      limit:
        maximum: 1000
        minimum: 1
        type: integer
      page:
        minimum: 1
        type: integer
      setting_key:
        type: string
      store_id:
        type: integer
    required:
    - limit
    - page
    type: object
  internal_modules_setting.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  internal_modules_setting.UpdateRequest:
    properties:
      description:
        maxLength: 255
        type: string
      id:
        type: integer
      setting_key:
        maxLength: 30
        minLength: 2
        type: string
      setting_values:
        type: string
      store_id:
        type: integer
    required:
    - id
    - setting_key
    - setting_values
    type: object
  internal_modules_upload.FileDeleteRequest:
    properties:
      id:
        type: integer
      store_id:
        type: integer
    required:
    - id
    type: object
  internal_modules_upload.FileListRequest:
    properties:
      file_name:
        type: string
      file_type:
        type: integer
      group_id:
        type: string
      limit:
        maximum: 100
        minimum: 1
        type: integer
      page:
        minimum: 1
        type: integer
      store_id:
        type: integer
    required:
    - limit
    - page
    type: object
  internal_modules_upload.GroupCreateRequest:
    properties:
      name:
        maxLength: 50
        minLength: 2
        type: string
      parent_id:
        type: integer
      sort:
        type: integer
    required:
    - name
    type: object
  internal_modules_upload.GroupDeleteRequest:
    properties:
      id:
        type: integer
      store_id:
        type: integer
    required:
    - id
    type: object
  internal_modules_upload.GroupUpdateRequest:
    properties:
      id:
        type: integer
      name:
        maxLength: 50
        minLength: 2
        type: string
      parent_id:
        type: integer
      sort:
        type: integer
    required:
    - id
    - name
    type: object
  internal_modules_upload.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      cost:
        type: string
      data: {}
      errcode:
        type: integer
      message:
        type: string
      traceId:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  title: Zero API
  version: "1.0"
paths:
...
  /setting/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.CreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建设置
      tags:
      - 设置管理
  /setting/default/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.DefaultCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建默认设置
      tags:
      - 默认设置管理
  /setting/default/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.DefaultDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除默认设置
      tags:
      - 默认设置管理
  /setting/default/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.DefaultListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_setting.ListResult'
              type: object
      summary: 获取默认设置列表
      tags:
      - 默认设置管理
  /setting/default/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.DefaultUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新默认设置
      tags:
      - 默认设置管理
  /setting/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.DeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除设置
      tags:
      - 设置管理
  /setting/form-configs:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.FormConfigsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取表单配置
      tags:
      - 设置管理
  /setting/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.ListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_setting.ListResult'
              type: object
      summary: 获取设置列表
      tags:
      - 设置管理
...
  /setting/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_setting.UpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新设置
      tags:
      - 设置管理
  /upload/file/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_upload.FileDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除文件
      tags:
      - 文件管理
  /upload/file/list:
    post:
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_upload.FileListRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_upload.ListResult'
              type: object
      summary: 获取文件列表
      tags:
      - 文件管理
  /upload/file/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传文件
        in: formData
        name: file
        required: true
        type: file
      - description: 分组ID
        in: formData
        name: group_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 上传文件(管理端)
      tags:
      - 文件管理
  /upload/group/create:
    post:
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_upload.GroupCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建上传分组
      tags:
      - 上传分组管理
  /upload/group/delete:
    post:
      parameters:
      - description: 删除参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_upload.GroupDeleteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除上传分组
      tags:
      - 上传分组管理
...
  /upload/group/update:
    post:
      parameters:
      - description: 更新参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_upload.GroupUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新上传分组
      tags:
      - 上传分组管理
...
swagger: "2.0"